rename some files for case consistency
