+++ packages/components/table-document/src/componentTypes.ts
