+++ packages/framework/framework-definitions/src/clipboard/componentClipboardProvider.ts
