+++ packages/framework/framework-definitions/src/clipboard/index.ts
@@ -2,5 +2,5 @@
  * Copyright (c) Microsoft Corporation. All rights reserved.
  * Licensed under the MIT License.
  */
-export * from "./IComponentClipboardProvider";
-export * from "./IComponentClipboardConsumer";
+export * from "./componentClipboardProvider";
+export * from "./componentClipboardConsumer";
