+++ packages/framework/framework-definitions/src/index.ts
@@ -7,4 +7,4 @@ export * from "./discover";
 export * from "./collection";
 export * from "./layout";
 export * from "./clipboard";
-export * from "./IComponentDefaultFactoryName";
+export * from "./componentDefaultFactoryName";
